requirements.txt
PyMuPDF>=1.23.0
google-genai>=0.3.0
python-dotenv>=1.0.0
tqdm>=4.66.0
httpx>=0.25.0